+++ lib/enums/navbar_enums.dart
@@ -1,0 +1,11 @@
+enum SimpleNavAnimType {
+  circle, // انیمیشن دایره‌ای
+  zoom, // انیمیشن زوم
+  float, // انیمیشن پرش (Float)
+}
+
+enum SimpleNavTextMode {
+  alwaysShow, // همیشه نمایش بده
+  onSelect, // فقط وقتی انتخاب شد
+  neverShow, // هرگز نمایش نده
+}