+++ lib/src/models/nav_bar_item.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class SimpleNavBarItem {
+  /// The text label displayed based on settings.
+  final String label;
+
+  /// The icon data to display.
+  final IconData icon;
+
+  SimpleNavBarItem({
+    required this.label,
+    required this.icon,
+  });
+}